pyproject.toml
[project]
name = "comfy-openwebuiapi"
description = "Add nodes for interacting with Ollama through the Open WebUI API."
version = "1.0.3"
license = {file = "LICENSE"} 
# classifiers = [
#     # For OS-independent nodes (works on all operating systems)
#     "Operating System :: OS Independent",
#
#     # OR for OS-specific nodes, specify the supported systems:
#     "Operating System :: Microsoft :: Windows",  # Windows specific
#     "Operating System :: POSIX :: Linux",  # Linux specific
#     "Operating System :: MacOS",  # macOS specific
#
#     # GPU Accelerator support. Pick the ones that are supported by your extension.
#     "Environment :: GPU :: NVIDIA CUDA",    # NVIDIA CUDA support
#     "Environment :: GPU :: AMD ROCm",       # AMD ROCm support
#     "Environment :: GPU :: Intel Arc",      # Intel Arc support
#     "Environment :: NPU :: Huawei Ascend",  # Huawei Ascend support
#     "Environment :: GPU :: Apple Metal",    # Apple Metal support
# ]

dependencies = ["pillow", "requests", "torch", "numpy"]

[project.urls]
Repository = "https://github.com/Gladon4/comfy-openwebuiapi"
#  Used by Comfy Registry https://registry.comfy.org
Documentation = "https://github.com/Gladon4/comfy-openwebuiapi/wiki"
"Bug Tracker" = "https://github.com/Gladon4/comfy-openwebuiapi/issues"

[tool.comfy]
PublisherId = "grille"
DisplayName = "Open WebUI API"
Icon = ""
includes = [] 
# "requires-comfyui" = ">=1.0.0"  # ComfyUI version compatibility

